MED; DeEntitize
Former-commit-id: 4e8c62305b0129edd6306bed0b43b6a95d5bc92e [formerly 3195b9e47f232394f844fa9fd0630c2ddd4cd400]
Former-commit-id: 41a80a852ae9b4963384b2e66c4e11f5cfced01b

+++ ExClient/Galleries/Metadata/LanguageName.cs
@@ -76,6 +76,6 @@
         Ukrainian,
         /// <summary>越南语</summary>
         Vietnamese,
-        Other = ushort.MaxValue
+        Other = int.MaxValue
     }
 }
